restrucutre ui helper funktions
