+++ src/ui/config_window/mod.rs
@@ -8,17 +8,18 @@ use crate::dataloading::dataprovider::song_data_provider::{
 use crate::dataloading::songinfo::SongInfo;
 use crate::ui::config_window::sidebar::{Sidebar, SidebarMessage};
 use crate::ui::widget::dynamic_text_input::DynamicTextInput;
-use crate::ui::{material_icon, material_icon_sized, with_tooltip};
+use crate::ui::widgets::buttons::material_symbol_message_button;
+use crate::ui::widgets::{material_symbol, material_symbol_sized};
+use crate::ui::with_tooltip;
 use crate::{DanceInterpreter, Message, Window};
 use fuzzy_matcher::FuzzyMatcher;
 use fuzzy_matcher::skim::SkimMatcherV2 as Matcher;
 use iced::alignment::Vertical;
 use iced::widget::{
-    Button, Column, Container, Row, Scrollable, Space, TextInput, button, checkbox, column as col,
-    container, radio, row, scrollable, text, toggler,
+    Column, Container, Row, Scrollable, Space, TextInput, button, column as col, container, row,
+    scrollable, text,
 };
-use iced::{Alignment, Element, Length, Pixels, Renderer, Size, Theme, window};
-use iced_aw::iced_aw_font;
+use iced::{Alignment, Element, Length, Size, Theme, window};
 use std::sync::LazyLock;
 use std::time::Instant;
 
@@ -106,7 +107,7 @@ impl ConfigWindow {
 
     fn build_search_bar<'a>(&self) -> Row<'a, Message> {
         row![
-            material_icon("search", false)
+            material_symbol("search", false)
                 .width(Length::Fixed(24.0))
                 .align_y(Vertical::Center),
             TextInput::new("Search...", &self.search_query)
@@ -114,8 +115,8 @@ impl ConfigWindow {
                 .on_submit(Message::Noop)
                 .width(Length::Fill)
                 .padding([4, 8]),
-            material_icon_message_button("backspace", false, Message::ClearSearch),
-            material_icon_message_button("close", false, Message::ToggleSearch),
+            material_symbol_message_button("backspace", false, Message::ClearSearch),
+            material_symbol_message_button("close", false, Message::ToggleSearch),
         ]
         .spacing(5)
         .padding([5, 5])
@@ -167,7 +168,7 @@ impl ConfigWindow {
             text("Load Playlist").size(20),
             button(
                 col![
-                    material_icon_sized("folder_open", false, 64),
+                    material_symbol_sized("folder_open", false, 64),
                     text("Open playlist file").size(16)
                 ]
                 .align_x(Alignment::Center)
@@ -183,7 +184,7 @@ impl ConfigWindow {
             text("Use Traktor").size(20),
             button(
                 col![
-                    material_icon_sized("agriculture", false, 64),
+                    material_symbol_sized("agriculture", false, 64),
                     text("Open the sidebar").size(16)
                 ]
                 .align_x(Alignment::Center)
@@ -217,7 +218,7 @@ impl ConfigWindow {
     fn build_statics_view(&'_ self, dance_interpreter: &DanceInterpreter) -> Column<'_, Message> {
         let trow = row![
             text("Statics").size(24).width(Length::Fill),
-            button(row![material_icon("add", false), text("Add Static")].spacing(5))
+            button(row![material_symbol("add", false), text("Add Static")].spacing(5))
                 .on_press(Message::AddBlankStatic)
                 .padding([5, 10])
                 .style(button::primary),
@@ -246,7 +247,7 @@ impl ConfigWindow {
 
     fn build_static_row<'a>(song: &SongInfo, idx: usize) -> Row<'a, Message> {
         row![
-            material_icon_message_button(
+            material_symbol_message_button(
                 "star",
                 song.is_favorite,
                 Message::ToggleStaticFavorite(idx)
@@ -257,7 +258,7 @@ impl ConfigWindow {
             row![
                 Space::new().width(Length::Fill).height(Length::Shrink),
                 with_tooltip(
-                    material_icon_message_button(
+                    material_symbol_message_button(
                         "smart_display",
                         false,
                         Message::SongChanged(SongChange::StaticAbsolute(idx))
@@ -265,7 +266,7 @@ impl ConfigWindow {
                     "Show now"
                 ),
                 with_tooltip(
-                    material_icon_message_button("delete", false, Message::DeleteStatic(idx)),
+                    material_symbol_message_button("delete", false, Message::DeleteStatic(idx)),
                     "Delete static"
                 ),
             ]
@@ -285,19 +286,19 @@ impl ConfigWindow {
         let (is_current, is_next, is_traktor, is_played) =
             dance_interpreter.data_provider.get_play_state(idx);
         let icon: Element<Message> = if is_traktor {
-            material_icon("agriculture", false)
+            material_symbol("agriculture", false)
                 .width(Length::Fixed(24.0))
                 .into()
         } else if is_current {
-            material_icon("play_arrow", false)
+            material_symbol("play_arrow", false)
                 .width(Length::Fixed(24.0))
                 .into()
         } else if is_next {
-            material_icon("skip_next", false)
+            material_symbol("skip_next", false)
                 .width(Length::Fixed(24.0))
                 .into()
         } else if is_played {
-            material_icon("check", false)
+            material_symbol("check", false)
                 .width(Length::Fixed(24.0))
                 .into()
         } else {
@@ -321,7 +322,7 @@ impl ConfigWindow {
             row![
                 Space::new().width(Length::Fill).height(Length::Shrink),
                 with_tooltip(
-                    material_icon_message_button(
+                    material_symbol_message_button(
                         "smart_display",
                         false,
                         Message::SongChanged(SongChange::PlaylistAbsolute(idx))
@@ -329,7 +330,7 @@ impl ConfigWindow {
                     "Show now"
                 ),
                 with_tooltip(
-                    material_icon_message_button(
+                    material_symbol_message_button(
                         "queue_play_next",
                         false,
                         Message::SetNextSong(SongDataSource::Playlist(idx))
@@ -337,7 +338,7 @@ impl ConfigWindow {
                     "Set as next song"
                 ),
                 with_tooltip(
-                    material_icon_message_button(
+                    material_symbol_message_button(
                         "delete",
                         false,
                         Message::DeleteSong(SongDataSource::Playlist(idx))
@@ -410,150 +411,4 @@ fn empty_folder_button_style(theme: &Theme, status: button::Status) -> button::S
     style.background = Some(palette.background.weakest.color.into());
     style.border.radius = 12.0.into();
     style
-}
-
-fn label_message_button_fill<'a>(
-    label: impl text::IntoFragment<'a>,
-    message: Message,
-) -> Button<'a, Message> {
-    label_message_button(label, message).width(Length::Fill)
-}
-
-fn label_message_button_shrink<'a>(
-    label: impl text::IntoFragment<'a>,
-    message: Message,
-) -> Button<'a, Message> {
-    label_message_button(label, message).width(Length::Shrink)
-}
-
-fn label_message_button<'a>(
-    label: impl text::IntoFragment<'a>,
-    message: Message,
-) -> Button<'a, Message> {
-    button(text(label).align_y(Vertical::Center))
-        .padding([4, 8])
-        .style(button::secondary)
-        .on_press(message)
-}
-
-#[allow(dead_code)]
-fn submenu_button(label: &'_ str) -> Button<'_, Message, Theme, Renderer> {
-    button(
-        row![
-            text(label).width(Length::Fill).align_y(Vertical::Center),
-            iced_aw_font::right_open()
-                .width(Length::Shrink)
-                .align_y(Vertical::Center),
-        ]
-        .align_y(Alignment::Center),
-    )
-    .padding([4, 8])
-    .style(button::text)
-    .on_press(Message::Noop)
-    .width(Length::Fill)
-}
-
-fn label_message_button_opt(label: &'_ str, message: Option<Message>) -> Button<'_, Message> {
-    if let Some(message) = message {
-        label_message_button(label, message)
-    } else {
-        button(text(label).align_y(Vertical::Center))
-            .padding([4, 8])
-            .style(button::primary)
-    }
-}
-
-fn label_message_button_fill_opt(label: &'_ str, message: Option<Message>) -> Button<'_, Message> {
-    label_message_button_opt(label, message).width(Length::Fill)
-}
-
-fn material_icon_message_button(
-    icon_id: &'_ str,
-    filled: bool,
-    message: Message,
-) -> Button<'_, Message> {
-    button(material_icon(icon_id, filled))
-        //.padding([4, 8])
-        .style(button::secondary)
-        .on_press(message)
-        .width(Length::Shrink)
-}
-
-fn material_icon_sized_message_button(
-    icon_id: &'_ str,
-    filled: bool,
-    size: impl Into<Pixels>,
-    message: Message,
-) -> Button<'_, Message> {
-    button(material_icon_sized(icon_id, filled, size))
-        .style(button::secondary)
-        .on_press(message)
-        .width(Length::Shrink)
-}
-
-fn labeled_message_checkbox(
-    label: &'_ str,
-    checked: bool,
-    message: fn(bool) -> Message,
-) -> checkbox::Checkbox<'_, Message> {
-    checkbox(checked)
-        .label(label)
-        .on_toggle(message)
-        .width(Length::Fill)
-    //.style(checkbox::secondary)
-}
-
-fn labeled_message_toggler(
-    label: &'_ str,
-    checked: bool,
-    message: fn(bool) -> Message,
-) -> toggler::Toggler<'_, Message> {
-    toggler(checked)
-        .label(label)
-        .on_toggle(message)
-        .width(Length::Fill)
-}
-
-#[allow(dead_code)]
-fn labeled_message_radio<T: Copy + Eq>(
-    label: &'_ str,
-    value: T,
-    selection: T,
-    message: fn(T) -> Message,
-) -> radio::Radio<'_, Message> {
-    radio(label, value, Some(selection), message).width(Length::Fill)
-    //.style(checkbox::secondary)
-}
-
-#[allow(dead_code)]
-fn labeled_message_checkbox_opt(
-    label: &'_ str,
-    checked: bool,
-    message: Option<fn(bool) -> Message>,
-) -> checkbox::Checkbox<'_, Message> {
-    if let Some(message) = message {
-        labeled_message_checkbox(label, checked, message)
-    } else {
-        checkbox(checked).label(label).width(Length::Fill)
-        //.style(checkbox::secondary)
-    }
-}
-
-#[allow(dead_code)]
-fn labeled_dynamic_text_input<'a>(
-    label: &'a str,
-    placeholder: &'a str,
-    value: &'a str,
-    message: fn(String) -> Message,
-    submit_message: Option<Message>,
-) -> Column<'a, Message> {
-    let mut input = DynamicTextInput::<Message>::new(placeholder, value)
-        .width(Length::Fill)
-        .on_change(message);
-
-    if let Some(submit_message) = submit_message {
-        input = input.on_submit(submit_message);
-    }
-
-    col!(text(label).width(Length::Fill), input,).width(Length::Fill)
 }
